+++ core/src/banking_stage/transaction_scheduler/transaction_priority_id.rs
@@ -1,8 +1,4 @@
-use {
-    crate::banking_stage::scheduler_messages::TransactionId,
-    prio_graph::TopLevelId,
-    std::hash::{Hash, Hasher},
-};
+use crate::banking_stage::scheduler_messages::TransactionId;
 
 /// A unique identifier tied with priority ordering for a transaction/packet:
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
@@ -14,18 +10,6 @@ pub(crate) struct TransactionPriorityId {
 impl TransactionPriorityId {
     pub(crate) fn new(priority: u64, id: TransactionId) -> Self {
         Self { priority, id }
-    }
-}
-
-impl Hash for TransactionPriorityId {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.id.hash(state)
-    }
-}
-
-impl TopLevelId<Self> for TransactionPriorityId {
-    fn id(&self) -> Self {
-        *self
     }
 }